refactor: constant apply
